handle armor toughness as float + scale toughness by 2 in toughness bar
closes #56

+++ core/src/main/java/dev/cheos/armorpointspp/core/adapter/IDataProvider.java
@@ -3,8 +3,8 @@ package dev.cheos.armorpointspp.core.adapter;
 public interface IDataProvider {
 	int armor();
 	int maxArmor();
-	int toughness();
-	int maxToughness();
+	double toughness();
+	double maxToughness();
 	int guiTicks();
 	int invulnTime();
 	long millis();
